refactor: data flow, hooks, types etc
