+++ apps/dapp/src/components/vaults/TitleBar/index.tsx
@@ -3,32 +3,33 @@ import { useEffect, useMemo } from 'react';
 import { Menu } from '@dopex-io/ui';
 
 import useVaultQuery from 'hooks/vaults/query';
-import useVaultState, { DurationType } from 'hooks/vaults/state';
+import useVaultStore from 'hooks/vaults/useVaultStore';
 
 import TitleItem from 'components/vaults/TitleBar/TitleItem';
 
 import { formatAmount } from 'utils/general';
 
-import { VAULTS_MENU } from 'constants/vaults/dropdowns';
+import { MARKETS_MENU } from 'constants/vaults/markets';
 
 interface Props {
-  selectedToken?: string;
-  handleSelectToken: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  market?: string;
+  handleSelectMarket: (e: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
 const TitleBar = (props: Props) => {
-  const { selectedToken = 'wstETH', handleSelectToken } = props;
-  const update = useVaultState((state) => state.update);
-  const vault = useVaultState((state) => state.vault);
+  const { market = 'ARB', handleSelectMarket } = props;
+
+  const update = useVaultStore((state) => state.update);
+  const vault = useVaultStore((state) => state.vault);
+
   const { vaults, aggregatedStats } = useVaultQuery({
-    vaultSymbol: selectedToken,
+    vaultSymbol: market,
   });
 
   const selectedVault = useMemo(() => {
     const selected = vaults.find(
       (_vault) =>
-        vault.durationType === _vault.durationType &&
-        vault.isPut === _vault.isPut
+        vault.duration === _vault.duration && vault.isPut === _vault.isPut
     );
 
     return selected;
@@ -38,27 +39,27 @@ const TitleBar = (props: Props) => {
     if (!selectedVault) return;
     update({
       base: selectedVault.underlyingSymbol,
-      isPut: selectedVault.isPut,
-      durationType: selectedVault.durationType as DurationType,
+      duration: selectedVault.duration,
       currentEpoch: selectedVault.currentEpoch,
       address: selectedVault.contractAddress,
       underlyingPrice: aggregatedStats?.currentPrice || 0,
+      isPut: selectedVault.isPut,
     });
   }, [selectedVault, update, aggregatedStats]);
 
   return (
     <div className="flex space-x-4 mb-4">
       <img
-        src={`/images/tokens/${selectedToken.toLowerCase()}.svg`}
+        src={`/images/tokens/${market.toLowerCase()}.svg`}
         className="w-[32px] h-[32px] my-auto border rounded-full border-carbon"
-        alt={selectedToken}
+        alt={market}
       />
       <Menu
         color="mineshaft"
         dropdownVariant="icon"
-        handleSelection={handleSelectToken}
-        selection={selectedToken}
-        data={VAULTS_MENU}
+        handleSelection={handleSelectMarket}
+        selection={market}
+        data={MARKETS_MENU}
         className="z-20"
         showArrow
       />
