+++ apps/dapp/src/types/ssov.ts
@@ -1,0 +1,21 @@
+import { Address } from 'viem';
+
+export const ssovDurations = ['WEEKLY', 'MONTHLY'] as const;
+
+export type SsovDuration = (typeof ssovDurations)[number];
+
+interface Vault {
+  symbol: string;
+  isPut: boolean;
+  duration: SsovDuration;
+  underlyingSymbol: string;
+  address: Address;
+}
+
+export interface SsovMarket {
+  vaults: Vault[];
+  default: {
+    isPut: boolean;
+    duration: SsovDuration;
+  };
+}
