+++ apps/dapp/src/utils/ssov/findSsov.ts
@@ -1,0 +1,21 @@
+import { SsovDuration } from 'types/ssov';
+
+import MARKETS from 'constants/vaults/markets';
+
+const findSsov = (
+  marketName: string,
+  isPut: boolean,
+  duration: SsovDuration
+) => {
+  const market = MARKETS[marketName];
+
+  if (!market) return;
+
+  return market.vaults.find((vault) => {
+    if (vault.duration === duration && vault.isPut === isPut) {
+      return true;
+    }
+  });
+};
+
+export default findSsov;
